remove k8s config refactoring
