+++ receiver/singletonreceivercreator/factory.go
@@ -3,7 +3,7 @@ package singletonreceivercreator
 import (
 	"context"
 
-	"github.com/kyma-project/opentelemetry-collector-components/internal/k8sconfig"
+	"github.com/kyma-project/opentelemetry-collector-components/receiver/singletonreceivercreator/internal/k8sconfig"
 
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/consumer"
