+++ receiver/singletonreceivercreator/internal/k8sconfig/config.go
@@ -2,10 +2,6 @@ package k8sconfig
 
 import (
 	"fmt"
-	"os"
-	"path/filepath"
-
-	"k8s.io/client-go/tools/clientcmd"
 
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
@@ -22,36 +18,16 @@ const (
 	// AuthTypeServiceAccount means to use the built-in service account that
 	// K8s automatically provisions for each pod.
 	AuthTypeServiceAccount AuthType = "serviceAccount"
-	// AuthTypeKubeConfig uses local credentials like those used by kubectl.
-	AuthTypeKubeConfig AuthType = "kubeconfig"
 )
 
-var authTypes = map[AuthType]bool{
-	AuthTypeServiceAccount: true,
-	AuthTypeKubeConfig:     true,
-}
-
 func GetK8sClient(authType AuthType) (kubernetes.Interface, error) {
-	if !authTypes[authType] {
+	if authType != AuthTypeServiceAccount {
 		return nil, fmt.Errorf("authentication type: %s not supported", string(authType))
 	}
-	var config *rest.Config
-	var err error
-	switch authType {
-	case AuthTypeServiceAccount:
-		config, err = rest.InClusterConfig()
-		if err != nil {
-			return nil, err
-		}
-	case AuthTypeKubeConfig:
-		kubeConfigPath := os.Getenv("KUBECONFIG")
-		if kubeConfigPath == "" {
-			kubeConfigPath = filepath.Join(os.Getenv("HOME"), ".kube/config")
-		}
-		config, err = clientcmd.BuildConfigFromFlags("", kubeConfigPath)
-		if err != nil {
-			return nil, err
-		}
+
+	config, err := rest.InClusterConfig()
+	if err != nil {
+		return nil, err
 	}
 
 	client, err := kubernetes.NewForConfig(config)
